backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-extensions==3.2.3
django-model-utils==5.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.12.2
